Remove DataStore class.
It wasn't reallly useful, and landuse is now stored in the Region class
anyways, so it is easier just to store a list of the values of the region class
for each timestep. I just need to make sure I make a deep copy of the class so
the stored ones aren't mutable.