Finish validate_hazard function.
Finished writing the validate_hazard function. Also added a new
validate_nseq_float function that was required for validating the bounds to
which hazards apply, and reworked the validate_int function to throw an error
if a float is passed to it. Previously the validate_int function would quietly
coerce floats to ints.