Fix rcsetup to update params from originalvalues.
The code was having problems when updating from an RC file, as it was trying to
update the rcParams instance with values that were no longer strings (as they
had already been validated and converted).

Also some other changes are made for clarity in the code, and several
deprecated comments were removed.