Major changes - csv results, speed and more.
Update code to write out a csv result file in the model results directory. This
csv file can be used for easy plotting in R or any other package. This change
also required adding month and year tracking to the Results object in
eventtracking.py.

Also, added 'speed' output to the ChitwanABMrc file generated in the model
output. The speed is just the number of seconds the model run took for each
model month. Lower speed values mean mean the model is running faster.