Read NBH coords from separate CSV file.
Don't use the coordinates accompanying the ICPSR dataset as they are in the
wrong coordinate system. The ICPSR dataset uses UTM45N, whereas all my data is
in UTM44N. So read the coords from a CSV file that has them in UTM44N instead
of using the ICPSR coordinates.