Fix to store NAs for events in time 0.
The code added in the past few commits allows storing the model initialization
in the time "zero" position (the zeroth timestep). Given that the model has not
started yet on the 0th timestep, the number of births, marriages, and other
life events is not known for time zero. This code ensures that number of events
is stored as NaN (not a number) for month 0, which, when read into the R code
as NAs, keeps them from being plotted as zeros.