Use hostname as part of model run ID.
This will allow merging of runs for the same scenario from different computers even if two runs started at the same time on the two different computers. Also, no longer pring the machine hostname on its own line int he ChitwanABMrc file output with each model run - it is now printed anyways as it is on the run ID line.